backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.14
pyjwt==2.8.0
python-dotenv==1.0.0
psycopg2-binary==2.9.11
passlib[bcrypt]==1.7.4
cohere>=4.0.0
email-validator==2.1.0
